+++ androidApp/src/androidMain/kotlin/at/asitplus/wallet/app/android/NdefDeviceEngagementService.kt
@@ -9,6 +9,7 @@ import android.os.VibrationEffect
 import android.os.Vibrator
 import androidx.core.content.ContextCompat
 import at.asitplus.catchingUnwrapped
+import at.asitplus.wallet.app.common.BuildType
 import at.asitplus.wallet.app.common.DummyPlatformAdapter
 import at.asitplus.wallet.app.common.ErrorService
 import at.asitplus.wallet.app.common.WalletConfig
@@ -269,7 +270,8 @@ class NdefDeviceEngagementService : HostApduService() {
                 dataStore = getDataStore(applicationContext),
                 platformAdapter = DummyPlatformAdapter()
             ),
-            errorService = ErrorService(serviceScope)
+            errorService = ErrorService(serviceScope),
+            buildType = BuildType.valueOf(BuildConfig.BUILD_TYPE.uppercase()),
         )
         serviceErrorService = walletConfig.errorService
         activeServiceErrorService = serviceErrorService